Update LangFormat.h
Need to store the members in their own struct. This would enable "multi-parsing" where a class structure could be determined from multiple places within the JSON doc, effectively merging different uses of a similar object and preventing types not being able to be parse from objects where some values are null or not "type-able". Structures should instead be hashed by their member names (this could cause clashes when 2 classes have identical member names [This is stupid anyway and we are not idiot proof]) and similar definitions can be merged if the current iteration is incomplete